user cRUD

+++ src/main/java/com/realestate/manager/repository/RoleRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 
-public interface RoleRepository extends JpaRepository<Role, Long>{
+public interface RoleRepository extends JpaRepository<Role, Integer>{
     Optional<Role> findByRoleName(String name);
 }
